admin et employeur done

+++ kergui-service-app/src/app/admin-employeur/publication-employeur/publication-employeur.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Profession } from 'src/app/models/profession';
 import { CandidatureServiceService } from 'src/app/services/candidature-service.service';
+import { ProfessionServiceService } from 'src/app/services/profession-service.service';
 import { PublicationService } from 'src/app/services/publication.service';
 import Swal from 'sweetalert2';
 
@@ -32,12 +33,15 @@ export class PublicationEmployeurComponent implements OnInit {
       created_at:'',
       updated_at:'',
     }
+publication: any;
 
     
 
-  constructor( private publicationservice : PublicationService, public candidatservice : CandidatureServiceService) { }
+  constructor( private publicationservice : PublicationService, private candidatservice : CandidatureServiceService, private professionservice : ProfessionServiceService) { }
   ngOnInit(): void {
   this.afficherPublication();
+  this.listeProfession ();
+
   }
   dataOffres : any;
   afficherOfrreUser(): void{
@@ -63,13 +67,28 @@ export class PublicationEmployeurComponent implements OnInit {
 
   userId = localStorage.getItem('userId');
   dataPublication : any;
+  userConnect : any;
   afficherPublication () : void{
-    this.publicationservice.geyAllpublication().subscribe((data)=>{
-      this.dataPublication = data.data;
-      // console.log("Tous les publication", this.dataPublication);
-      // stocker les annones dans local storage
-      // localStorage.setItem('annonce', JSON.stringify(data))
-    })
+
+    const recupuserConnecte = localStorage.getItem('user_profile')
+    this.userConnect = recupuserConnecte ? JSON.parse(recupuserConnecte) : null;
+    if(this.userConnect){
+      
+      this.publicationservice.geyAllpublication().subscribe((data)=>{
+        this.dataPublication = data.data;
+        console.log("dddd", this.dataPublication);
+        
+
+        this.dataPublication.forEach((publication: {  user_id: any; }) => {
+          if(publication.user_id === this.userConnect.id){
+            // console.log("Publication de l'utilisateur connecté :", publication);
+            // console.log("Informations de l'utilisateur :", this.userConnect);
+          }
+        });
+      })
+      
+    }
+
   }
 
 
@@ -91,6 +110,8 @@ ajouterPublicatin (): void{
   this.publicationservice.addPubication(this.newPublication).subscribe((datapubli : any) =>{
     console.log("champ", this.newPublication);
     console.log("ajout bien", datapubli);
+
+    this.viderChamp();
     
     window.location.reload();
   },
@@ -99,19 +120,18 @@ ajouterPublicatin (): void{
   }
   )
 }
+// methode pour vider champ
+viderChamp(): void{
+  this.lieu = "";
+  this.typeContrat = "";
+  this.description = "";
+  this.slaireMinimum = "";
+  this.experienceMinimum  = "";
+  this.profession_id = "";
+  
+}
 
 selectedPublication : any;
-
-// lieu:"",
-// typeContrat:"",
-// description:"",
-// slaireMinimum:"",
-// experienceMinimum:"",
-// Profession_id:"",
-// user_id: '',
-// profession_id:'',
-// created_at:'',
-// updated_at:'',
 
 // methode pour charger publication à modifier
 chargerPublication( publication : any){
@@ -147,37 +167,40 @@ chargerPublication( publication : any){
       cancelButtonColor: "#d33",
       confirmButtonText: "Oui modifier!"
     }).then((result) => {
-      if (this.selectedPublication) {
-        // Modification d'une profession existante
-        this.publicationservice.editPublication(this.selectedPublication, {
-          lieu: this.lieu,
-    typeContrat : this.typeContrat,
-    description : this.description,
-    slaireMinimum : this.slaireMinimum,
-    experienceMinimum :  this.experienceMinimum,
-    userId : this.userId,
-    profession_id : this.profession_id,
-        
-        }).subscribe(
-          (data: any) => {
-            console.log("Modification réussie :", data);
-            // Effectuez les actions nécessaires après la modification, par exemple, actualiser la liste
-            // window.location.reload(); ou mieux, mettre à jour la liste localement
-            
-            this.afficherPublication();  // mettre a jour la liste
-          },
-          error => {
-            console.error('Erreur lors de la modification :', error);
-          }
-        );
-      } else {
-        console.error("Erreur: Aucune profession sélectionnée pour la modification");
-        // Gérez l'erreur ou fournissez un message à l'utilisateur si nécessaire
-        Swal.fire({
-          title: "service modifié!",
-          text: "Ce service a été modifié .",
-          icon: "success"
-          });
+      if(result.isConfirmed){
+
+        if (this.selectedPublication) {
+          // Modification d'une profession existante
+          this.publicationservice.editPublication(this.selectedPublication, {
+            lieu: this.lieu,
+      typeContrat : this.typeContrat,
+      description : this.description,
+      slaireMinimum : this.slaireMinimum,
+      experienceMinimum :  this.experienceMinimum,
+      userId : this.userId,
+      profession_id : this.profession_id,
+          
+          }).subscribe(
+            (data: any) => {
+              console.log("Modification réussie :", data);
+              // Effectuez les actions nécessaires après la modification, par exemple, actualiser la liste
+              // window.location.reload(); ou mieux, mettre à jour la liste localement
+              
+              this.afficherPublication();  // mettre a jour la liste
+            },
+            error => {
+              console.error('Erreur lors de la modification :', error);
+            }
+          );
+        } else {
+          console.error("Erreur: Aucune profession sélectionnée pour la modification");
+          // Gérez l'erreur ou fournissez un message à l'utilisateur si nécessaire
+          Swal.fire({
+            title: "publication modifié!",
+            text: "Cet publicaion  a été supprimé.",
+            icon: "success"
+            });
+        }
       }
       }
     )
@@ -186,21 +209,50 @@ chargerPublication( publication : any){
 // methode pour supprimer publication
 // methode pour supprmer un element
 supprimerProfession(id: any): void {
-  this.publicationservice.deletePublication(id).subscribe(
-    (reponse: any) => {
-      console.log("supprimer success", reponse);
-      // Mettez à jour la liste locale des professions après la suppression
-      this.afficherPublication();
-    },
-    error => {
-      console.error(`Erreur lors de la suppression de la profession avec l'ID ${id} :`, error);
-    }
-  );
+  
+  Swal.fire({
+    title: "Voulez vous vraiment supprime cette publication?",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#FF9A00",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Oui supprimer!"
+  }).then((result) => { 
+    if(result.isConfirmed){
+
+      this.publicationservice.deletePublication(id).subscribe(
+        (reponse: any) => {
+          console.log("supprimer success", reponse);
+          // Mettez à jour la liste locale des professions après la suppression
+          this.afficherPublication();
+        },
+        error => {
+          console.error(`Erreur lors de la suppression de la profession avec l'ID ${id} :`, error);
+          Swal.fire({
+            title: "publication supprime!",
+            text: "Cette publication  a été supprimé .",
+            icon: "success"
+            });
+        }
+        
+      );
+    } 
+   })
+
+ 
 }
 
 publcationselect : any;
 detailPublication(element : any){
   this.publcationselect =element;
+}
+// methode pour recupererer tous les profession
+professionData : any;
+listeProfession () :  void{
+  this.professionservice.getProfession().subscribe((data)=>{
+this.professionData = data.data;
+console.log("voir profession", this.professionData);
+  })
 }
 
 }
